# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Voice as a Service",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/clients": {
   "get": {
    "tags": [
     "clients"
    ],
    "summary": "List Clients",
    "description": "Admin: todos. Usuario de un cliente: el suyo.",
    "operationId": "list_clients_api_v1_clients_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ClientOut"
         },
         "type": "array",
         "title": "Response List Clients Api V1 Clients Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "clients"
    ],
    "summary": "Create Client",
    "operationId": "create_client_api_v1_clients_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ClientIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClientOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/clients/{client_id}": {
   "get": {
    "tags": [
     "clients"
    ],
    "summary": "Read Client",
    "operationId": "read_client_api_v1_clients__client_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "client_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Client Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClientOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "clients"
    ],
    "summary": "Update Client",
    "operationId": "update_client_api_v1_clients__client_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "client_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Client Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ClientUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClientOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "clients"
    ],
    "summary": "Delete Client",
    "description": "Solo sin conversaciones; con historial, desactivarlo (active=false).",
    "operationId": "delete_client_api_v1_clients__client_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "client_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Client Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/clients/{client_id}/usage": {
   "get": {
    "tags": [
     "clients"
    ],
    "summary": "Client Usage",
    "description": "Consumo del mes contra los limites del tier.",
    "operationId": "client_usage_api_v1_clients__client_id__usage_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "client_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Client Id"
      }
     },
     {
      "name": "month",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
        },
        {
         "type": "null"
        }
       ],
       "description": "YYYY-MM; sin el, el mes en curso",
       "title": "Month"
      },
      "description": "YYYY-MM; sin el, el mes en curso"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UsageOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ClientIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Name"
     },
     "slug": {
      "type": "string",
      "pattern": "^[a-z0-9][a-z0-9_]{0,63}$",
      "title": "Slug"
     },
     "tier_id": {
      "type": "string",
      "title": "Tier Id"
     },
     "active": {
      "type": "boolean",
      "title": "Active",
      "default": true
     },
     "retention_days": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Retention Days",
      "description": "Dias que se guardan las conversaciones (null: los del tier)"
     }
    },
    "type": "object",
    "required": [
     "name",
     "slug",
     "tier_id"
    ],
    "title": "ClientIn"
   },
   "ClientOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "active": {
      "type": "boolean",
      "title": "Active"
     },
     "tier": {
      "$ref": "#/components/schemas/TierBrief"
     },
     "retention_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Retention Days"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "agents_count": {
      "type": "integer",
      "title": "Agents Count",
      "default": 0
     },
     "numbers_count": {
      "type": "integer",
      "title": "Numbers Count",
      "default": 0
     },
     "effective_retention_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Effective Retention Days"
     },
     "effective_max_call_seconds": {
      "type": "integer",
      "title": "Effective Max Call Seconds",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "slug",
     "active",
     "tier",
     "retention_days",
     "created_at"
    ],
    "title": "ClientOut"
   },
   "ClientUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 128,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "tier_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tier Id"
     },
     "active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Active"
     },
     "retention_days": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Retention Days"
     }
    },
    "type": "object",
    "title": "ClientUpdate"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MinutesUsageOut": {
    "properties": {
     "used_seconds": {
      "type": "integer",
      "title": "Used Seconds"
     },
     "used_minutes": {
      "type": "number",
      "title": "Used Minutes"
     },
     "limit_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Limit Minutes"
     },
     "remaining_minutes": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Remaining Minutes"
     }
    },
    "type": "object",
    "required": [
     "used_seconds",
     "used_minutes",
     "limit_minutes",
     "remaining_minutes"
    ],
    "title": "MinutesUsageOut"
   },
   "NumbersUsageOut": {
    "properties": {
     "used": {
      "type": "integer",
      "title": "Used"
     },
     "limit": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Limit"
     }
    },
    "type": "object",
    "required": [
     "used",
     "limit"
    ],
    "title": "NumbersUsageOut"
   },
...
   "TierBrief": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "TierBrief"
   },
...
   "UsageOut": {
    "properties": {
     "month": {
      "type": "string",
      "title": "Month"
     },
     "period_start": {
      "type": "string",
      "format": "date-time",
      "title": "Period Start"
     },
     "period_end": {
      "type": "string",
      "format": "date-time",
      "title": "Period End"
     },
     "active_calls": {
      "type": "integer",
      "title": "Active Calls"
     },
     "max_concurrent_calls": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Concurrent Calls"
     },
     "inbound": {
      "$ref": "#/components/schemas/MinutesUsageOut"
     },
     "outbound": {
      "$ref": "#/components/schemas/MinutesUsageOut"
     },
     "phone_numbers": {
      "$ref": "#/components/schemas/NumbersUsageOut"
     }
    },
    "type": "object",
    "required": [
     "month",
     "period_start",
     "period_end",
     "active_calls",
     "max_concurrent_calls",
     "inbound",
     "outbound",
     "phone_numbers"
    ],
    "title": "UsageOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "Token de la sesion de demo o de la llamada",
    "scheme": "bearer"
   }
  }
 }
}